+++ lib/cztop.rb
@@ -31,6 +31,7 @@ require_relative 'cztop/config/traversing'
 require_relative 'cztop/config/serialization'
 require_relative 'cztop/message/frames'
 require_relative 'cztop/socket/types'
+require_relative 'cztop/z85/padded'
 
 
 # make Ctrl-C work in case a low-level call hangs